mod op

+++ src/operators/tensor/math/modulo.cairo
@@ -59,7 +59,7 @@ fn modulo<
     divisor.shape.len().print();
 
     if dividend.shape.len() < divisor.shape.len() {
-        let mut i: usize = 0;
+        let mut i: i32 = 0;
         loop {
             if i >= dividend.shape.len() {
                 break;
@@ -83,7 +83,7 @@ fn modulo<
     }
 
     if dividend.shape.len() > divisor.shape.len() {
-        let mut i: usize = 0;
+        let mut i: i32 = 0;
         loop {
             if i >= divisor.shape.len() {
                 break;
